--- controllers/userControllers.js
@@ -1,21 +1,0 @@
-
-
-const showDashboard = (req, res) => {
-    try {
-        res.render('userViews/dashboard')
-    } catch (error) {
-        console.log('FAILED to RENDER dashboard')
-    }
-}
-
-const showSearch = (req, res) => {
-    try {
-        res.render('userViews/search')
-    } catch (error) {
-        console.log('FAILED to RENDER search')
-    }
-}
-module.exports = {
-    showDashboard,
-    showSearch
-}
