+++ routers/frontRouters.js
@@ -1,17 +1,27 @@
 const express = require("express");
 const router = express.Router();
-const { searchTitle, getIndex, getFavouriteMovies, getMovie, searchMovie
+const { searchTitle, getIndex, getFavouriteMovies, getMovie, searchMovie, showDashboard, showSearch
 } = require("../controllers/frontControllers");
 
-router.get("/search-title/:title", searchTitle);
-router.get("/movies", getFavouriteMovies);
-
+//INDEX renderView
 router.get("/", getIndex);
 
-//router.get('/search', getMovie)
-//router.post("/search", getMovie); 
-router.get('/search', searchMovie)
-router.post("/search", searchMovie); 
+//DASHBOARD renderView
+router.get("/dashboard", showDashboard);
+
+//SEARCH renderView
+router.get("/search", showSearch)
+
+//Buscar pelis por título, devuelve una sola.
+router.get("/search-title/:title", searchTitle);
+
+//myMovies renderView
+router.get("/myMovies", getFavouriteMovies);
+
+router.get('/found-all', getMovie);
+router.post("/found-all", getMovie);
+// router.get('/found', searchMovie);
+// router.post('/found', searchMovie);
