+++ views/404.ejs
@@ -3,7 +3,7 @@
 <main>
     <div id="block404">
         <figure>
-            <img id="404Img" src="./assets/groot404.png" alt="grootSadPNG">
+            <img id="404Img" src="\assets\groot404.png" alt="grootSadPNG">
         </figure>
     </div>
     <div>
