+++ views/userViews/myMovies.ejs
@@ -1,0 +1,30 @@
+<%- include('./userTemplates/header2') %> <%- include("userTemplates/nav") %> <%
+if (arrayMovies.length > 0) { %>
+<h1 class="centerText">My Movies</h1>
+<section class="favoutie-movies gridContainerMax-2c">
+  <% arrayMovies.forEach(movie => { %>
+  <article class="movie-card centerText gridContainerMax-2c maxWidthMedium">
+    <div>
+      <img src="<%=movie.Poster%>" alt="<%=movie.Title %> " />
+      <a href= <%=`/search/${movie.Title}`%> >See More</a>
+    </div>
+    <div>
+      <h1><%=movie.Title %></h1>
+      <div class="movie-facts">
+        <span>Year: <%=movie.Year%>, </span
+        ><span>Duration: <%= movie.Runtime %>, </span>
+        <span>Genre: <%= movie.Genre %> </span>
+      </div>
+      <h4>Director:</h4>
+      <p><%= movie.Director %></p>
+      <a href="#">Remove from favourites</a>
+    </div>
+  </article>
+  <% }) %>
+</section>
+<% } else { %>
+<p class="error-message">No movies found</p>
+<% } %> <%- include("./userTemplates/footer") %>
+
+
+
