+++ views/userViews/search.ejs
@@ -2,7 +2,16 @@
 <%- include("./userTemplates/photoSliderOpacity.ejs") %>
 
 <main class="centerText">
-    aqui va una barra de busqueda dinámina en CSS
+    <form action="http://localhost:3000/found-all" method="POST">
+        <input type="text" name="search" id="search" placeholder="Ingresa el título de la película">
+        <button type="submit">Buscar</button>
+    </form>
 </main>
+
+      <!-- <form action="http://localhost:3000/found-all" method="POST">
+        <label for="search">Search</label>
+        <input id="search" name="search" type="search" pattern=".*\S.*" required>
+        <span class="caret"></span>
+    </form> -->
 
 <%- include("./userTemplates/footer") %>
